% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_check.R
\name{extract_spectra_matrix}
\alias{extract_spectra_matrix}
\title{Extract spectra matrix and pixel metadata from a Cardinal MSI object}
\usage{
extract_spectra_matrix(msi_obj)
}
\arguments{
\item{msi_obj}{A Cardinal MSI object imported by \code{Cardinal::readImzML()}.}
}
\value{
A list containing:
\describe{
\item{spectra}{Numeric matrix with pixels as rows and m/z features as columns.}
\item{pixel_info}{A data.frame of pixel metadata.}
\item{mz}{Numeric vector of m/z values.}
\item{msi_obj}{The MSI object with ensured \code{pixel_ID} in \code{pixelData}.}
}
}
\description{
Extract spectra matrix and pixel metadata from a Cardinal MSI object
}
