% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{minmax_scale}
\alias{minmax_scale}
\title{Min-max scale a matrix by columns}
\usage{
minmax_scale(x)
}
\arguments{
\item{x}{A numeric matrix with pixels/samples as rows and features as columns.}
}
\value{
A numeric matrix of the same dimension as \code{x}.
}
\description{
Perform column-wise min-max scaling:
}
\details{
\deqn{(x - min(x)) / (max(x) - min(x))}

Each feature is rescaled independently to the range \code{[0,1]}.
}
