% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{run_clustering}
\alias{run_clustering}
\title{Run clustering on embedding coordinates}
\usage{
run_clustering(embedding, method = "kmeans", centers = 10L)
}
\arguments{
\item{embedding}{A data.frame or matrix containing embedding coordinates.
Typically the output of \code{run_umap_py()}.}

\item{method}{Character string specifying the clustering method.
Currently only \code{"kmeans"} is supported.}

\item{centers}{Integer; number of clusters. Used when \code{method = "kmeans"}.
Default is \code{10L}.}
}
\value{
A list containing clustering results. For \code{method = "kmeans"},
the returned list includes:
\describe{
\item{method}{The clustering method used.}
\item{result}{The raw \code{stats::kmeans()} result object.}
\item{cluster}{A factor vector of cluster assignments.}
}
}
\description{
This function performs clustering on low-dimensional embedding coordinates,
such as UMAP results. Currently, only \code{"kmeans"} clustering is supported,
but additional methods may be added in future versions.
}
\details{
For \code{method = "kmeans"}, clustering is performed on all columns of the
embedding coordinates, and the random seed is fixed internally at
\code{2026} for reproducibility.
}
\examples{
emb <- data.frame(UMAP1 = rnorm(20), UMAP2 = rnorm(20))
res <- run_clustering(emb, method = "kmeans", centers = 3)
table(res$cluster)

}
